requirements.txt
aiofiles==23.1.0
altair==5.0.1
anyio==3.6.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
decorator==5.1.1
ecdsa==0.18.0
extra-streamlit-components==0.1.56
fastapi==0.95.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.0
packaging==23.1
pandas==2.0.3
passlib==1.7.4
Pillow==9.5.0
protobuf==4.23.3
psycopg2-binary==2.9.6
pyarrow==12.0.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMySQL==1.0.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
rich==13.4.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
streamlit==1.24.1
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
websockets==10.4
zipp==3.15.0
